# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deer Culling Records API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Stalker {
  id: number;
  name: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface CreateStalkerRequest {
  name: string;
  pin: string;
  isAdmin?: boolean;
}

export interface UpdateStalkerRequest {
  name?: string;
  pin?: string;
  isAdmin?: boolean;
}

export interface LoginRequest {
  name: string;
  pin: string;
}

export type CullRecordSpecies =
  (typeof CullRecordSpecies)[keyof typeof CullRecordSpecies];

export const CullRecordSpecies = {
  red_deer: "red_deer",
  roe_deer: "roe_deer",
  fallow_deer: "fallow_deer",
  sika_deer: "sika_deer",
  muntjac: "muntjac",
  chinese_water_deer: "chinese_water_deer",
} as const;

export type CullRecordSex = (typeof CullRecordSex)[keyof typeof CullRecordSex];

export const CullRecordSex = {
  stag: "stag",
  hind: "hind",
  buck: "buck",
  doe: "doe",
} as const;

export type CullRecordCondition =
  (typeof CullRecordCondition)[keyof typeof CullRecordCondition];

export const CullRecordCondition = {
  excellent: "excellent",
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface CullRecord {
  id: number;
  stalkerId?: number | null;
  stalkerName?: string | null;
  species: CullRecordSpecies;
  sex: CullRecordSex;
  weight?: number | null;
  condition: CullRecordCondition;
  pregnant?: boolean | null;
  latitude: number;
  longitude: number;
  notes?: string | null;
  culledAt: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateCullRequestSpecies =
  (typeof CreateCullRequestSpecies)[keyof typeof CreateCullRequestSpecies];

export const CreateCullRequestSpecies = {
  red_deer: "red_deer",
  roe_deer: "roe_deer",
  fallow_deer: "fallow_deer",
  sika_deer: "sika_deer",
  muntjac: "muntjac",
  chinese_water_deer: "chinese_water_deer",
} as const;

export type CreateCullRequestSex =
  (typeof CreateCullRequestSex)[keyof typeof CreateCullRequestSex];

export const CreateCullRequestSex = {
  stag: "stag",
  hind: "hind",
  buck: "buck",
  doe: "doe",
} as const;

export type CreateCullRequestCondition =
  (typeof CreateCullRequestCondition)[keyof typeof CreateCullRequestCondition];

export const CreateCullRequestCondition = {
  excellent: "excellent",
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface CreateCullRequest {
  stalkerId?: number | null;
  species: CreateCullRequestSpecies;
  sex: CreateCullRequestSex;
  weight?: number | null;
  condition: CreateCullRequestCondition;
  pregnant?: boolean | null;
  latitude: number;
  longitude: number;
  notes?: string | null;
  culledAt: string;
}

export type UpdateCullRequestSpecies =
  (typeof UpdateCullRequestSpecies)[keyof typeof UpdateCullRequestSpecies];

export const UpdateCullRequestSpecies = {
  red_deer: "red_deer",
  roe_deer: "roe_deer",
  fallow_deer: "fallow_deer",
  sika_deer: "sika_deer",
  muntjac: "muntjac",
  chinese_water_deer: "chinese_water_deer",
} as const;

export type UpdateCullRequestSex =
  (typeof UpdateCullRequestSex)[keyof typeof UpdateCullRequestSex];

export const UpdateCullRequestSex = {
  stag: "stag",
  hind: "hind",
  buck: "buck",
  doe: "doe",
} as const;

export type UpdateCullRequestCondition =
  (typeof UpdateCullRequestCondition)[keyof typeof UpdateCullRequestCondition];

export const UpdateCullRequestCondition = {
  excellent: "excellent",
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface UpdateCullRequest {
  stalkerId?: number | null;
  species?: UpdateCullRequestSpecies;
  sex?: UpdateCullRequestSex;
  weight?: number | null;
  condition?: UpdateCullRequestCondition;
  pregnant?: boolean | null;
  latitude?: number;
  longitude?: number;
  notes?: string | null;
  culledAt?: string;
}

export type ListCullsParams = {
  stalkerId?: number;
  /**
   * Cull-season start year (May–April), e.g. 2026 for 2026/27
   */
  season?: number;
};