# lib/api-zod/src/generated/types/stalker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deer Culling Records API
 * OpenAPI spec version: 0.2.0
 */

export interface Stalker {
  id: number;
  name: string;
  isAdmin: boolean;
  createdAt: Date;
}